Profile section improved - but orders not pulling through.

+++ checkout/views.py
@@ -7,6 +7,8 @@ from django.conf import settings
 import stripe
 from shopping_bag.contexts import shopping_bag_contents
 from products.models import Product
+from profiles.forms import UserProfileForm
+from profiles.models import UserProfile
 from .forms import OrderForm
 from .models import Order, OrderLineItem
 import json
@@ -33,7 +35,7 @@ def cache_checkout_data(request):
 def checkout(request):
     stripe_public_key = settings.STRIPE_PUBLIC_KEY
     stripe_secret_key = settings.STRIPE_SECRET_KEY
-    client_secret = None
+    # client_secret = None   #  Line added by Marcel
 
     if request.method == 'POST':
         print("Request Method = POST!!")
@@ -52,48 +54,48 @@ def checkout(request):
         }
         order_form = OrderForm(form_data)
         print("Checking if order form is valid.....")
-        # if order_form.is_valid():
-        print("Order form is validated!! :-)")
-        order = order_form.save(commit=False)
-        pid = request.POST.get('client_secret').split('_secret')[0]
-        order.stripe_pid = pid
-        order.save()
-        order.original_shopping_bag = json.dumps(shopping_bag_session)
-        for item_id, item_data in shopping_bag_session.items():
-            try:
-                product = Product.objects.get(id=item_id)
-                if isinstance(item_data, int):
-                    order_line_item = OrderLineItem(
-                        order=order,
-                        product=product,
-                        quantity=item_data,
-                    )
-                    order_line_item.save()
-                    print("order_line_item = ", order_line_item)
-                else:
-                    for size, quantity in item_data['items_by_size'].items():
+        if order_form.is_valid():
+            print("Order form is validated!! :-)")
+            order = order_form.save(commit=False)
+            pid = request.POST.get('client_secret').split('_secret')[0]
+            order.stripe_pid = pid
+            order.save()
+            order.original_shopping_bag = json.dumps(shopping_bag_session)
+            for item_id, item_data in shopping_bag_session.items():
+                try:
+                    product = Product.objects.get(id=item_id)
+                    if isinstance(item_data, int):
                         order_line_item = OrderLineItem(
                             order=order,
                             product=product,
-                            quantity=quantity,
-                            product_size=size,
+                            quantity=item_data,
                         )
                         order_line_item.save()
                         print("order_line_item = ", order_line_item)
-            except Product.DoesNotExist:
-                messages.error(request, (
-                    "One of the products in your bag wasn't found in our database. "
-                    "Please call us for assistance!")
-                )
-                order.delete()
-                return redirect(reverse('shopping_bag'))
-        print("Order Line Items added - attempting to render 'Checkout Success'")
-        request.session['save_info'] = 'save-info' in request.POST
-        return redirect(reverse('checkout_success', args=[order.order_number]))
-        # else:
-        #     print("Order for not validated - it is invalid!! :-(")
-        #     messages.error(request, 'There was an error with your form. \
-        #         Please double check your information.')
+                    else:
+                        for size, quantity in item_data['items_by_size'].items():
+                            order_line_item = OrderLineItem(
+                                order=order,
+                                product=product,
+                                quantity=quantity,
+                                product_size=size,
+                            )
+                            order_line_item.save()
+                            print("order_line_item = ", order_line_item)
+                except Product.DoesNotExist:
+                    messages.error(request, (
+                        "One of the products in your bag wasn't found in our database. "
+                        "Please call us for assistance!")
+                    )
+                    order.delete()
+                    return redirect(reverse('shopping_bag'))
+            print("Order Line Items added - attempting to render 'Checkout Success'")
+            request.session['save_info'] = 'save-info' in request.POST
+            return redirect(reverse('checkout_success', args=[order.order_number]))
+        else:
+            print("Order for not validated - it is invalid!! :-(")
+            messages.error(request, 'There was an error with your form. \
+                Please double check your information.')
 
     else:
         print("Request Method is 'GET' ???")
@@ -112,7 +114,7 @@ def checkout(request):
             amount=stripe_total,
             currency=settings.STRIPE_CURRENCY
         )
-        client_secret = intent.client_secret
+        # client_secret = intent.client_secret     #  Line added by Marcel
 
         order_form = OrderForm()
 
@@ -124,7 +126,8 @@ def checkout(request):
     context = {
         'order_form': order_form,
         'stripe_public_key': stripe_public_key,
-        'client_secret': client_secret,
+        # 'client_secret': client_secret,     #  Line 'amended' by Marcel to replace the line below.
+        'client_secret': intent.client_secret,
     }
 
     return render(request, template, context)
@@ -136,6 +139,29 @@ def checkout_success(request, order_number):
     """
     save_info = request.session.get('save_info')
     order = get_object_or_404(Order, order_number=order_number)
+
+    if request.user.is_authenticated:
+       
+        profile = UserProfile.objects.get(user=request.user)
+        # Attach the user's profile to the order
+        order.user_profile = profile
+        order.save()
+
+        # Save the user's info
+        if save_info:
+            profile_data = {
+                'default_phone_number': order.phone_number,
+                'default_country': order.country,
+                'default_postcode': order.postcode,
+                'default_town_or_city': order.town_or_city,
+                'default_street_address1': order.street_address1,
+                'default_street_address2': order.street_address2,
+                'default_county': order.county,
+            }
+            user_profile_form = UserProfileForm(profile_data, instance=profile)
+            if user_profile_form.is_valid():
+                user_profile_form.save()
+
     messages.success(request, f'Order successfully processed! \
         Your order number is {order_number}. A confirmation \
         email will be sent to {order.email}.')
